Adding new world bosses support with 1.1 API

+++ Modules/Ao/Taraviza.php
@@ -1,6 +1,6 @@
 <?php
 /*
-* Tara.php - Handle Tara timer/alert.
+* Tara.php - Handle Tara timer/alert (plus season's event world bosses)
 * With linkage to Boss/Buff Timers API courtesy of The Nadybot Team
 * BeBot - An Anarchy Online & Age of Conan Chat Automaton
 * Copyright (C) 2004 Jonas Jax
@@ -14,7 +14,7 @@
 * - Khalem (RK1)
 * - Naturalistic (RK1)
 * - Temar (RK1)
-* - Bitnykk (RK2) for Gauntlet addition
+* - Bitnykk (RK2) for Gauntlet/WorldBosses additions
 * See Credits file for all aknowledgements.
 *
 *  This program is free software; you can redistribute it and/or modify
@@ -53,6 +53,12 @@ class Taraviza extends BaseActiveModule
 		$this -> register_command("all", "settara", "GUEST");
 		$this -> register_command("all", "setviza", "GUEST");
 		$this -> register_command("all", "gauntbuff", "GUEST");
+		$this -> register_command("all", "world", "GUEST");
+		$this -> register_alias("world", "worldbosses");
+		$this -> register_alias("world", "worldboss");
+		$this -> register_alias("world", "wboss");
+		$this -> register_alias("world", "wbosses");
+		$this -> register_alias("world", "wb");		
 		$this -> register_alias("gauntbuff", "gaunbuff");
 		$this -> register_alias("gauntbuff", "gaubuff");
 		$this -> register_alias("gauntbuff", "gbuff");		
@@ -64,31 +70,34 @@ class Taraviza extends BaseActiveModule
 		$this -> register_alias("tara", "spawn");
 		$this -> register_alias("tara", "pop");
 		$this -> register_alias("tara", "poptime");
-		$this -> help['description'] = 'Allows to have a specific Tarasque/Gauntlet timer. Both auto-update by API every 6 hours but is still manually overwritable.';
+		$this -> help['description'] = 'Allows to have a specific Tarasque/Gauntlet timers & warnings ; both auto-update by API every 6 hours but are still manually overwritable. Also added seasonal !world bosses that are fully automated & could pop more randomly hence no alert/manual option.';
 		$this -> help['command']['tara']="Shows time left to next Tarasque.";
 		$this -> help['command']['viza']="Shows time left to next Gauntlet.";
 		$this -> help['command']['gbuff']="Shows current state for Gauntlet's Buff.";
 		$this -> help['command']['settara'] = "Sets the timer from now. Add -/+ number for minute adjusting.";
 		$this -> help['command']['setviza'] = "Sets the timer from now. Add -/+ number for minute adjusting.";
+		$this -> help['command']['world'] = "Shows all bosses including those available at current season (could be anniversary, halloween, winter, other, etc).";
 		
 		$this->bot->core("settings")->create('Taraviza', 'TaraAlert', 'None', 'Towards which channel(s) should Tarasque pop alert be sent to (None = disable) ?', 'Both;Guildchat;Private;None');
 		$this->bot->core("settings")->create('Taraviza', 'VizaAlert', 'None', 'Towards which channel(s) should Gauntlet start alert be sent to(None = disable) ?', 'Both;Guildchat;Private;None');
-		$this->bot->core("settings")->create('Taraviza', 'PopAlertTime', 59, 'How long in minutes before pop/start should alerts be sent to selected channel(s)?', '14;23;32;41;50;59');
+		$this->bot->core("settings")->create('Taraviza', 'PopAlertTime', 59, 'How long in minutes before pop/start should non-seasonal alerts be sent to selected channel(s)?', '14;23;32;41;50;59');
         $this->bot->core("settings")
-            ->create("Taraviza", "AlertDisc", false, "Do we alert Discord of worldbosses spawns ?");
+            ->create("Taraviza", "AlertDisc", false, "Do we alert Discord of non-seasonal bosses spawns ?");
         $this->bot->core("settings")
-            ->create("Taraviza", "DiscChanId", "", "What Discord ChannelId in case we separate worldbosses spawns from main Discord channel (leave empty for all in main channel) ?");
+            ->create("Taraviza", "DiscChanId", "", "What Discord ChannelId in case we separate non-seasonal bosses spawns from main Discord channel (leave empty for all in main channel) ?");
         $this->bot->core("settings")
-            ->create("Taraviza", "DiscTag", "", "Should we add a Discord Tag (e.g. @here or @everyone) to worldbosses spams for notifying Discord users (leave empty for no notification) ?");
+            ->create("Taraviza", "DiscTag", "", "Should we add a Discord Tag (e.g. @here or @everyone) to non-seasonal bosses spams for notifying Discord users (leave empty for no notification) ?");
         $this->bot->core("settings")
-            ->create("Taraviza", "AlertIrc", false, "Do we alert Irc of worldbosses spawns ?");		
+            ->create("Taraviza", "AlertIrc", false, "Do we alert Irc of non-seasonal bosses spawns ?");		
         $this->bot->core("settings")
             ->create("Taraviza", "ApiUrl", "https://timers.aobots.org/api/", "What's the Boss/Buff API URL we should use to auto-update (Nadybot's by default, leave empty to disable automation) ?");			
 		$this->register_event("cron", "1min");
-		$this->register_event("cron", "6hour");
-		$this->tcycle=34200; // 9H30 cycle
-		$this->vcycle=61620;  // 17H07 cycle
-		$this->apiver='v1.0';
+		$this->register_event("cron", "3hour");
+		$this->tcycle=34200; // 9H30 tara cycle (30=immortality)
+		$this->vcycle=61620;  // 17H07 viza cycle (7=immortality)
+		$this->apiver='v1.1';
+		$this->wlist= '';
+		$this->wcount= 0;
 	}
 
 	function command_handler($name, $msg, $channel)
@@ -111,6 +120,9 @@ class Taraviza extends BaseActiveModule
 			return $this -> set_viza($name, "add", $info, $channel);
 		else if (preg_match("/^setviza [-]([0-9]+)$/i", $msg, $info))
 			return $this -> set_viza($name, "rem", $info, $channel);
+		else if (preg_match("/^world$/i", $msg)) {
+			return $this->wcount." world boss(es) currently found : ".$this->bot->core("tools")->make_blob("click to view", $this->wlist);
+		}
 	}
 
     function cron($cron)
@@ -160,7 +172,7 @@ class Taraviza extends BaseActiveModule
 					}				
 				}
 			}		
-		} elseif ($cron == 21600) {
+		} elseif ($cron == 10800) {
 			if($this->bot->core("settings")->get("Taraviza", "ApiUrl")!='') {
 				$now = time(); $t = false; $v = false;
 				
@@ -174,28 +186,68 @@ class Taraviza extends BaseActiveModule
 				foreach ($take as $line){ $timer = $line[0]; }
 				if($now>$timer&&$now-$timer>$this->vcycle) $v = true;
 				
-				if($v||$t) {
-					$url = $this->bot->core("settings")->get("Taraviza", "ApiUrl")."/".$this->apiver."/"."bosses";
-					$content = $this->bot->core("tools")->get_site($url);	
-					if (!($content instanceof BotError)) {
-						if (strpos($content, '{"name":') !== false) {
-							$timers = json_decode($content);													
-							foreach($timers as $timer) {
-								if($t&&$timer->name=='tara'&&$timer->last_spawn>0) {							
-									$this -> bot -> db -> query("TRUNCATE TABLE tara");
-									$this -> bot -> db -> query("INSERT INTO tara (time) VALUES ('".$timer->last_spawn."')");
-								}
-								if($v&&$timer->name=='vizaresh'&&$timer->last_spawn>0) {
-									$this -> bot -> db -> query("TRUNCATE TABLE viza");
-									$this -> bot -> db -> query("INSERT INTO viza (time) VALUES ('".$timer->last_spawn."')");
-								}							
-							}												
-						}
+				$url = $this->bot->core("settings")->get("Taraviza", "ApiUrl")."/".$this->apiver."/"."bosses";
+				$content = $this->bot->core("tools")->get_site($url);	
+				if (!($content instanceof BotError)) {
+					if (strpos($content, '{"name":') !== false) {							
+						$timers = json_decode($content);		
+						$this->wlist= ''; $this->wcount= 0;							
+						foreach($timers as $timer) {
+							if($t&&$timer->name=='tara'&&$this->bot->dimension==$timer->dimension&&$timer->last_spawn>0) {							
+								$this -> bot -> db -> query("TRUNCATE TABLE tara");
+								$this -> bot -> db -> query("INSERT INTO tara (time) VALUES ('".$timer->last_spawn."')");
+							}
+							if($v&&$timer->name=='vizaresh'&&$this->bot->dimension==$timer->dimension&&$timer->last_spawn>0) {
+								$this -> bot -> db -> query("TRUNCATE TABLE viza");
+								$this -> bot -> db -> query("INSERT INTO viza (time) VALUES ('".$timer->last_spawn."')");
+							}
+							$this -> wcount++;
+							$boss = $timer->name;
+							switch($boss) {
+								case 'atma':
+									$cycle = 10800; // 3H cycle
+									break;										
+								case 'cerubin':
+									$cycle = 32400; // 9h cycle
+									break;
+									case 'tara':
+									$cycle = $this->tcycle;
+									break;
+									case 'vizaresh':
+									$cycle = $this->vcycle;
+									break;										
+								case 'loren':
+								case 'father':
+								case 'reaper':
+									$cycle = 33300; // 9H15 cycle (15=immortality)
+									break;																											
+								default:
+									$cycle = 21600; // 6H default cycle, assumed for abmouth/tam/zal & any other
+									break;
+							}
+							$this -> wlist .= '<br>'.ucfirst($boss).' (RK'.$timer->dimension.') : may pop in '.$this->nextpop($timer->last_spawn,$cycle);
+						}											
 					}
 				}
 			}
 		}
 	}
+	
+	function nextpop($timer,$cycle)
+	{
+        $now = time();
+        while ($timer <= $now) { $timer = $timer + $cycle; }
+        $left = $timer - $now;
+        $hour = floor($left/3600);
+        $left = $left - ($hour*3600);
+        $min = floor($left/60);
+        $sec = $left - ($min*60);
+        if ($sec < 10) { $sec = "0".$sec; }
+        if ($hour < 10) { $hour = "0".$hour; }
+        if ($min < 10) { $min = "0".$min; }
+		$msg = $hour."h".$min."m";
+		return $msg;
+    }	
 
 	function show_buff()
 	{
@@ -212,20 +264,20 @@ class Taraviza extends BaseActiveModule
 							$faction = $buff->faction;
 							$expires = $buff->expires;
 						} else {
-							$return = "No active sided buff detected for now.";
+							$return .= " No active sided buff detected for RK".$buff->dimension." ";
 						}
-					}
-					if($faction!=''&&$expires>$now) {
-						$left = $expires - $now;
-						$hour = floor($left/3600);
-						$left = $left - ($hour*3600);
-						$min = floor($left/60);
-						$sec = $left - ($min*60);
-						if ($sec < 10) { $sec = "0".$sec; }
-						if ($hour < 10) { $hour = "0".$hour; }
-						if ($min < 10) { $min = "0".$min; }
-						$return = "Current buff is ".$faction." & should expire in about ".$hour."h".$min."m";						
-					}
+						if($faction!=''&&$expires>$now) {
+							$left = $expires - $now;
+							$hour = floor($left/3600);
+							$left = $left - ($hour*3600);
+							$min = floor($left/60);
+							$sec = $left - ($min*60);
+							if ($sec < 10) { $sec = "0".$sec; }
+							if ($hour < 10) { $hour = "0".$hour; }
+							if ($min < 10) { $min = "0".$min; }
+							$return .= " Current RK".$buff->dimension." buff is ".$faction." and should expire in about ".$hour."h".$min."m ";						
+						}
+					}					
 				} else {
 					$return = "No buff info for now, may retry later on.";
 				}
